// flyteconsole/packages/zapp/console/src/components/Launch/LaunchForm/types.ts
export interface LaunchRoles {
  authRole?: Admin.IAuthRole;
  securityContext?: Core.ISecurityContext;
}

export interface LaunchRoleInputRef {
  getValue(): LaunchRoles;
  validate(): boolean;
}
// ...
export enum AuthRoleTypes {
  k8 = 'k8',
  IAM = 'IAM',
}